Refine cardio week history mock handling
