+++ components/screens/ProgressScreen.tsx
@@ -2,12 +2,12 @@ import { useEffect, useMemo, useState } from "react";
 
 import AppScreen from "../layouts/AppScreen";
 import type { TimeRange } from "../../src/types/progress";
-import type { CardioHistoryEntry, HistoryEntry, ProgressDomain } from "../progress/Progress.types";
+import type { HistoryEntry, ProgressDomain } from "../progress/Progress.types";
 import { CARDIO_WEEK_HISTORY_MOCK, PROGRESS_MOCK_SNAPSHOTS } from "./progress/MockData";
 import { TrendOverview } from "./progress/TrendOverview";
 import { KPI_COLORS, getEncouragement, getKpiFormatter } from "./progress/util";
 import { HistorySection } from "./progress/HistorySection";
-import { CardioWeekHistory, type CardioWeekHistoryDay } from "./progress/CardioWeekHistory";
+import { CardioWeekHistory, buildCardioWeekHistory, type CardioWeekHistoryDay } from "./progress/CardioWeekHistory";
 import { KpiTiles } from "./progress/KpiTiles";
 import { useAuth } from "../AuthContext";
 import { logger } from "../../utils/logging";
@@ -20,14 +20,7 @@ import { RangeSelector } from "./progress/RangeSelector";
 import { DOMAIN_LABELS, DOMAIN_OPTIONS, RANGE_LABELS, RANGE_OPTIONS } from "./progress/constants";
 import { useCardioProgressSnapshot, useStrengthHistory, useUserFirstName } from "./progress/hooks";
 
-const USE_CARDIO_WEEK_HISTORY_MOCK = true;
-
-type AggregatedTotals = {
-  calories?: number;
-  distance?: number;
-  steps?: number;
-  time?: number;
-};
+const USE_CARDIO_WEEK_MOCK = true;
 
 interface ProgressScreenProps {
   bottomBar?: React.ReactNode;
@@ -63,7 +56,7 @@ export function ProgressScreen({ bottomBar, onSelectRoutine }: ProgressScreenPro
       return [];
     }
 
-    if (USE_CARDIO_WEEK_HISTORY_MOCK) {
+    if (USE_CARDIO_WEEK_MOCK) {
       return CARDIO_WEEK_HISTORY_MOCK.days;
     }
 
@@ -170,130 +163,4 @@ export function ProgressScreen({ bottomBar, onSelectRoutine }: ProgressScreenPro
 }
 
 export default ProgressScreen;
-
-function buildCardioWeekHistory(entries: HistoryEntry[]): CardioWeekHistoryDay[] {
-  const cardioEntries = entries.filter((entry): entry is CardioHistoryEntry => entry.type === "cardio");
-  if (cardioEntries.length === 0) {
-    return [];
-  }
-
-  const grouped = cardioEntries.reduce<Map<string, { date: Date; workouts: CardioWeekHistoryDay["workouts"]; totals: AggregatedTotals; label?: string }>>(
-    (acc, entry) => {
-      const date = new Date(entry.date);
-      const key = date.toISOString().split("T")[0];
-      if (!acc.has(key)) {
-        acc.set(key, {
-          date,
-          workouts: [],
-          totals: {},
-          label: getWeekdayLabel(date),
-        });
-      }
-      const group = acc.get(key)!;
-      const workout: CardioWeekHistoryDay["workouts"][number] = {
-        id: entry.id,
-        type: entry.type,
-        name: entry.activity,
-        duration: entry.duration,
-        distance: entry.distance,
-        calories: entry.calories,
-        time: entry.time,
-        steps: entry.steps,
-      };
-      group.workouts.push(workout);
-
-      const calories = parseNumeric(entry.calories);
-      if (typeof calories === "number") {
-        group.totals.calories = (group.totals.calories ?? 0) + calories;
-      }
-
-      const distance = parseNumeric(entry.distance);
-      if (typeof distance === "number") {
-        group.totals.distance = (group.totals.distance ?? 0) + distance;
-      }
-
-      if (typeof entry.steps === "number") {
-        group.totals.steps = (group.totals.steps ?? 0) + entry.steps;
-      }
-
-      const minutes = parseDurationMinutes(entry.duration);
-      if (typeof minutes === "number") {
-        group.totals.time = (group.totals.time ?? 0) + minutes;
-      }
-
-      return acc;
-    },
-    new Map(),
-  );
-
-  return Array.from(grouped.values())
-    .sort((a, b) => b.date.getTime() - a.date.getTime())
-    .map(({ date, workouts, totals, label }) => {
-      const weekIndex = getWeekIndex(date);
-      const formattedTotals: CardioWeekHistoryDay["dailyTotals"] = {
-        calories: typeof totals.calories === "number" ? Math.round(totals.calories) : totals.calories,
-        distance: typeof totals.distance === "number" ? `${totals.distance.toFixed(1)} km` : totals.distance,
-        steps: typeof totals.steps === "number" ? Math.round(totals.steps) : totals.steps,
-        time: typeof totals.time === "number" ? formatMinutes(totals.time) : totals.time,
-      };
-
-      return {
-        key: date.toISOString().split("T")[0],
-        label,
-        weekIndex,
-        dateLabel: formatDateLabel(date),
-        dailyTotals: formattedTotals,
-        workouts,
-      } satisfies CardioWeekHistoryDay;
-    });
-}
-
-function parseNumeric(value?: string) {
-  if (!value) return undefined;
-  const match = value.match(/[\d,.]+/);
-  if (!match) return undefined;
-  const normalized = match[0].replace(/,/g, "");
-  const parsed = Number.parseFloat(normalized);
-  return Number.isNaN(parsed) ? undefined : parsed;
-}
-
-function parseDurationMinutes(value?: string) {
-  if (!value) return undefined;
-
-  if (/^\d{1,2}:\d{2}:\d{2}$/.test(value)) {
-    const [hours, minutes, seconds] = value.split(":").map(Number);
-    if ([hours, minutes, seconds].some((part) => Number.isNaN(part))) {
-      return undefined;
-    }
-    return hours * 60 + minutes + seconds / 60;
-  }
-
-  const hourMatch = value.match(/(\d+)\s*h/);
-  const minuteMatch = value.match(/(\d+)\s*m/);
-  const totalMinutes =
-    (hourMatch ? Number.parseInt(hourMatch[1], 10) * 60 : 0) + (minuteMatch ? Number.parseInt(minuteMatch[1], 10) : 0);
-  return Number.isNaN(totalMinutes) ? undefined : totalMinutes;
-}
-
-function formatMinutes(totalMinutes: number) {
-  const hours = Math.floor(totalMinutes / 60);
-  const minutes = Math.round(totalMinutes % 60);
-  if (hours > 0) {
-    return `${hours}h ${minutes}m`;
-  }
-  return `${minutes}m`;
-}
-
-function getWeekIndex(date: Date) {
-  const jsDay = date.getDay();
-  return jsDay === 0 ? 6 : jsDay - 1;
-}
-
-function getWeekdayLabel(date: Date) {
-  return date.toLocaleDateString(undefined, { weekday: "short" }).slice(0, 1).toUpperCase();
-}
-
-function formatDateLabel(date: Date) {
-  return date.toLocaleDateString(undefined, { weekday: "short", month: "short", day: "numeric" });
-}
